requirements.txt
asgiref==3.4.1
colorama==0.4.4
Django==3.2.9
django-ckeditor==6.2.0
django-environ==0.8.1
django-js-asset==1.2.2
django-registration-redux==2.9
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
document==1.0
kwargs==1.0.1
Menu==3.2.2
Pillow==8.4.0
pip==21.3.1
python-decouple==3.5
psycopg2-binary==2.9.1
pytz==2021.3
qrcode==7.3.1
setuptools==58.5.3
sqlparse==0.4.2
wheel==0.37.0
